src/main/webapp/app/entities/type/type-update.component.ts
import { computed, defineComponent, inject, ref, type Ref } from 'vue';
import { useI18n } from 'vue-i18n';
import { useRoute, useRouter } from 'vue-router';
import { useVuelidate } from '@vuelidate/core';

import TypeService from './type.service';
import { useValidation } from '@/shared/composables';
import { useAlertService } from '@/shared/alert/alert.service';

import { type IType, Type } from '@/shared/model/type.model';

export default defineComponent({
  compatConfig: { MODE: 3 },
  name: 'TypeUpdate',
  setup() {
    const typeService = inject('typeService', () => new TypeService());
    const alertService = inject('alertService', () => useAlertService(), true);

    const type: Ref<IType> = ref(new Type());
    const isSaving = ref(false);
    const currentLanguage = inject('currentLanguage', () => computed(() => navigator.language ?? 'en'), true);

    const route = useRoute();
    const router = useRouter();

    const previousState = () => router.go(-1);

    const retrieveType = async typeId => {
      try {
        const res = await typeService().find(typeId);
        type.value = res;
      } catch (error) {
        alertService.showHttpError(error.response);
      }
    };

    if (route.params?.typeId) {
      retrieveType(route.params.typeId);
    }

    const initRelationships = () => {};

    initRelationships();

    const { t: t$ } = useI18n();
    const validations = useValidation();
    const validationRules = {
      nameType: {},
      books: {},
    };
    const v$ = useVuelidate(validationRules, type as any);
    v$.value.$validate();

    return {
      typeService,
      alertService,
      type,
      previousState,
      isSaving,
      currentLanguage,
      v$,
      t$,
    };
  },
  created(): void {},
  methods: {
    save(): void {
      this.isSaving = true;
      if (this.type.id) {
        this.typeService()
          .update(this.type)
          .then(param => {
            this.isSaving = false;
            this.previousState();
            this.alertService.showInfo(this.t$('bookApp.type.updated', { param: param.id }));
          })
          .catch(error => {
            this.isSaving = false;
            this.alertService.showHttpError(error.response);
          });
      } else {
        this.typeService()
          .create(this.type)
          .then(param => {
            this.isSaving = false;
            this.previousState();
            this.alertService.showSuccess(this.t$('bookApp.type.created', { param: param.id }).toString());
          })
          .catch(error => {
            this.isSaving = false;
            this.alertService.showHttpError(error.response);
          });
      }
    },
  },
});
